restore serializer template
