+++ tests/benchmark/bm_utils.cpp
@@ -5,8 +5,7 @@ void load_HNSW_index(const char *path, VecSimIndex *hnsw_index) {
     // Load the index file, if it exists in the expected path.
     auto location = std::string(getenv("ROOT"));
     auto file_name = location + "/" + path;
-    auto serializer =
-        HNSWIndexSerializer<float, float>(reinterpret_cast<HNSWIndex<float, float> *>(hnsw_index));
+    auto serializer = HNSWIndexSerializer(reinterpret_cast<HNSWIndex<float, float> *>(hnsw_index));
     std::ifstream input(file_name, std::ios::binary);
     if (input.is_open()) {
         serializer.loadIndex(file_name);
